backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-search-results>=2.4.2
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0